// server/types/enums.ts
export enum HtmlTag {
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  P = 'p',
  A = 'a',
  Div = 'div',
}

export enum ComponentName {
  UpdatedOn = 'UpdatedOn',
  Media = 'Media',
  MediaGrid = 'MediaGrid',
  StatGrid = 'StatGrid',
  BulletedList = 'BulletedList',
  Link = 'Link',
}

export enum MediaType {
  Img = 'img',
  Video = 'video',
}

export enum SseEventType {
  Status = 'status',
  Complete = 'complete',
  Error = 'error',
}